refactor: remove legacy list components and replace with new implementation
