--- src/lib/exp/list/index.ts
@@ -1,2 +1,0 @@
-export * from './list';
-export * from './list-item';
