--- src/lib/exp/list/list-item/index.ts
@@ -1,11 +1,0 @@
-import { defineCustomElement } from '@tylertech/forge-core';
-import { ListItemComponentExp } from './list-item';
-
-export * from './list-item-adapter';
-export * from './list-item-constants';
-export * from './list-item-foundation';
-export * from './list-item';
-
-export function defineListItemComponentExp(): void {
-  defineCustomElement(ListItemComponentExp);
-}
