--- src/lib/exp/list/list-item/list-item.html
@@ -1,18 +1,0 @@
-<template>
-  <div class="forge-list-item" part="root">
-    <slot name="leading"></slot>
-    <slot name="avatar"></slot>
-    <div class="text-container" part="text-container">
-      <slot></slot>
-      <slot name="title"></slot>
-      <slot name="primary-text"></slot>
-      <slot name="subtitle"></slot>
-      <slot name="secondary-text"></slot>
-      <slot name="tertiary-text"></slot>
-      <slot name="tertiary-title"></slot>
-    </div>
-    <slot name="trailing"></slot>
-    <forge-state-layer exportparts="surface:state-layer__surface" target=":host"></forge-state-layer>
-    <forge-focus-indicator exportparts="indicator:focus-indicator__indicator" target=":host" inward></forge-focus-indicator>
-  </div>
-</template>
