--- src/lib/exp/list/list/list.html
@@ -1,5 +1,0 @@
-<template>
-  <div class="forge-list" part="root">
-    <slot></slot>
-  </div>
-</template>
