--- src/lib/list/build.json
@@ -1,4 +1,0 @@
-{
-  "$schema": "../../../node_modules/@tylertech/forge-cli/build-schema.json",
-  "extends": "../build.json"
-}
