--- src/lib/list/list-item/_variables.scss
@@ -1,38 +1,0 @@
-@use 'sass:math';
-
-$block-padding: 8px;
-$inline-padding: 16px;
-$drawer-inline-padding: #{math.div($inline-padding, 2)};
-
-/// The default height for a single line list item.
-$single-line-height: #{48px - $block-padding * 2} !default;
-
-/// The default height for a single line list item.
-$dense-single-line-height: #{32px - $block-padding * 2} !default;
-
-/// The default height for a two line list item.
-$dense-two-line-height: #{56px - $block-padding * 2} !default;
-
-/// The default height for a three line list item.
-$dense-three-line-height: #{72px - $block-padding * 2} !default;
-
-/// The default min-height for a single line list item in a drawer context.
-$drawer-single-line-height: 40px !default;
-
-/// The default height for a two-line list item.
-$two-line-height: #{72px - $block-padding * 2} !default;
-
-/// The default height for a three-line list item.
-$three-line-height: #{88px - $block-padding * 2} !default;
-
-/// The default height for the dense single-line list item
-$dense-height: #{32px - $block-padding * 2} !default;
-
-/// The default margin for list items when placed in a drawer
-$margin-drawer: 8px !default;
-
-/// The default margin-left for indented list items
-$indent: 56px !default;
-
-/// The default margin-left for indented list items when placed in a drawer
-$indent-drawer: $indent + $margin-drawer !default;
