+++ src/lib/list/list-item/list-item-adapter.ts
@@ -1,162 +1,121 @@
-import { addClass, getShadowElement, removeClass, requireParent, isDeepEqual, toggleClass } from '@tylertech/forge-core';
+import { getShadowElement, isDeepEqual, toggleAttribute } from '@tylertech/forge-core';
+import { replaceElement } from '../../core/utils/utils';
 import { BaseAdapter, IBaseAdapter } from '../../core/base/base-adapter';
-import { createUserInteractionListener } from '../../core/utils';
+import { FOCUS_INDICATOR_CONSTANTS, IFocusIndicatorComponent } from '../../focus-indicator';
+import { IStateLayerComponent, STATE_LAYER_CONSTANTS } from '../../state-layer';
 import { IListComponent } from '../list/list';
-import { LIST_CONSTANTS } from '../list/list-constants';
+import { ListComponentItemRole, LIST_CONSTANTS } from '../list/list-constants';
 import { IListItemComponent } from './list-item';
 import { LIST_ITEM_CONSTANTS } from './list-item-constants';
-import { ForgeRipple } from '../../ripple';
 
-export interface IListItemAdapter extends IBaseAdapter {
-  initializeAccessibility(): void;
-  getListItem(): IListItemComponent;
-  addListener(type: string, listener: (evt: Event) => void, options?: AddEventListenerOptions): void;
-  removeListener(type: string, listener: (evt: Event) => void): void;
-  createRipple(): any;
-  setStatic(value: boolean): void;
-  setTwoLine(value: boolean): void;
-  setThreeLine(value: boolean): void;
-  getLineCount(): number;
+export interface IListItemAdapter extends IBaseAdapter<IListItemComponent> {
+  initialize(): void;
+  setHref(href: string, target: string): void;
+  setHrefTarget(target: string): void;
+  setNonInteractive(value: boolean): void;
+  setDisabled(value: boolean): void;
   setActive(value: boolean): void;
-  setSelected(value: boolean): void;
-  tryToggleCheckboxRadio(value?: boolean): void;
-  setFocus(): void;
-  hasFocus(): boolean;
-  setRole(role: string): void;
-  setDisabled(disabled: boolean): void;
-  setDense(dense: boolean): void;
-  setIndented(indented: boolean): void;
-  setWrap(value: boolean): void;
   trySelect(value: unknown): boolean | null;
-  userInteractionListener(): ReturnType<typeof createUserInteractionListener>;
+  tryToggleSelectionControl(value?: boolean): void;
+  isFocused(): boolean;
+  setFocus(): void;
+  animateStateLayer(): void;
 }
 
 export class ListItemAdapter extends BaseAdapter<IListItemComponent> implements IListItemAdapter {
-  private _listItemElement: HTMLElement;
-  private _defaultSlot: HTMLSlotElement;
+  private _rootElement: HTMLElement | HTMLAnchorElement;
+  private _focusIndicatorElement: IFocusIndicatorComponent;
+  private _stateLayerElement: IStateLayerComponent;
 
   constructor(component: IListItemComponent) {
     super(component);
-    this._initialize();
+    this._rootElement = getShadowElement(component, LIST_ITEM_CONSTANTS.selectors.ROOT);
+    this._focusIndicatorElement = getShadowElement(component, FOCUS_INDICATOR_CONSTANTS.elementName) as IFocusIndicatorComponent;
+    this._stateLayerElement = getShadowElement(component, STATE_LAYER_CONSTANTS.elementName) as IStateLayerComponent;
   }
 
-  private _initialize(): void {
-    this._listItemElement = getShadowElement(this._component, LIST_ITEM_CONSTANTS.selectors.LIST_ITEM);
-    this._defaultSlot = getShadowElement(this._component, LIST_ITEM_CONSTANTS.selectors.DEFAULT_SLOT) as HTMLSlotElement;
-  }
-  
-  public initializeAccessibility(): void {
+  public initialize(): void {
+    // Attempt to inherit state for corresponding properties from our parent list
+    const list = this._getParentList();
+    if (list) {
+      this._inheritParentListProps(list);
+    }
+
+    this._rootElement.tabIndex = this._component.nonInteractive || this._component.disabled ? -1 : 0;
+
     if (!this._component.hasAttribute('role')) {
-      this._component.setAttribute('role', 'listitem');
+      this._setRole(list);
     }
   }
 
-  public getListItem(): IListItemComponent {
-    return this._component;
+  public setHref(href: string, target: string): void {
+    if (href) {
+      if (this._rootElement.tagName !== 'A') {
+        const anchor = this._createAnchorRootElement(target);
+        this._rootElement = replaceElement(this._rootElement, anchor);
+      }
+      (this._rootElement as HTMLAnchorElement).href = href;
+    } else if (this._rootElement.tagName === 'A') {
+      const defaultEl = this._createDefaultRootElement();
+      this._rootElement = replaceElement(this._rootElement, defaultEl);
+    }
   }
 
-  /**
-   * Adds an event listener to the `<forge-list-item>` host element.
-   * @param {string} type The event type.
-   * @param {Function} listener The event callback.
-   */
-  public addListener(type: string, listener: (evt: Event) => void, options?: AddEventListenerOptions): void {
-    this._listItemElement.addEventListener(type, listener, options);
+  public setHrefTarget(target: string): void {
+    if (this._rootElement.tagName === 'A') {
+      (this._rootElement as HTMLAnchorElement).target = target;
+    }
   }
 
-  /**
-   * Removes an event listener from the `<forge-list-item>` host element.
-   * @param {string} type The event type.
-   * @param {Function} listener The event callback.
-   */
-  public removeListener(type: string, listener: (evt: Event) => void): void {
-    this._listItemElement.removeEventListener(type, listener);
-  }
+  public setNonInteractive(value: boolean): void {
+    this._rootElement.tabIndex = value ? -1 : 0;
 
-  /**
-   * Creates a ripple instance on the list item.
-   * @returns {ForgeRipple}
-   */
-  public createRipple(): any {
-    return new ForgeRipple(this._listItemElement);
-  }
-
-  /**
-   * Toggles the static state of this list item.
-   * @param {boolean} value The static state.
-   */
-  public setStatic(value: boolean): void {
     if (value) {
-      this._listItemElement.tabIndex = -1;
-      addClass(LIST_ITEM_CONSTANTS.classes.STATIC, this._listItemElement);
+      this._focusIndicatorElement.remove();
+      this._stateLayerElement.remove();
     } else {
-      this._listItemElement.tabIndex = 0;
-      removeClass(LIST_ITEM_CONSTANTS.classes.STATIC, this._listItemElement);
+      this._rootElement.append(this._focusIndicatorElement, this._stateLayerElement);
     }
   }
 
-  /**
-   * Sets this list item as a two-line list item.
-   * @param {boolean} value The two-line state.
-   */
-  public setTwoLine(value: boolean): void {
+  public setDisabled(value: boolean): void {
+    this._rootElement.tabIndex = value ? -1 : 0;
+    toggleAttribute(this._component, value, 'aria-disabled', 'true');
+
     if (value) {
-      addClass(LIST_ITEM_CONSTANTS.classes.TWO_LINE, this._listItemElement);
+      this._focusIndicatorElement.remove();
+      this._stateLayerElement.remove();
     } else {
-      removeClass(LIST_ITEM_CONSTANTS.classes.TWO_LINE, this._listItemElement);
+      this._rootElement.append(this._focusIndicatorElement, this._stateLayerElement);
     }
   }
 
-  /**
-   * Sets this list item as a three-line list item.
-   * @param {boolean} value The three-line state.
-   */
-  public setThreeLine(value: boolean): void {
-    if (value) {
-      addClass(LIST_ITEM_CONSTANTS.classes.THREE_LINE, this._listItemElement);
-    } else {
-      removeClass(LIST_ITEM_CONSTANTS.classes.THREE_LINE, this._listItemElement);
-    }
-  }
-
-  /**
-   * Determines how many lines the list item is displaying.
-   */
-  public getLineCount(): number {
-    return this._defaultSlot.assignedNodes().filter(e => e.nodeType === Node.ELEMENT_NODE).length; // assignedElements does not exist in the polyfilled version so filtering nodes instead
-  }
-
-  /**
-   * Toggles the active class of the list item.
-   * @param {boolean} value The active state.
-   */
   public setActive(value: boolean): void {
-    if (value) {
-      addClass(LIST_ITEM_CONSTANTS.classes.ACTIVE, this._listItemElement);
-    } else {
-      removeClass(LIST_ITEM_CONSTANTS.classes.ACTIVE, this._listItemElement);
-    }
+    this._focusIndicatorElement.active = value;
   }
 
   /**
-   * Toggles the selected class of the list item.
-   * @param {boolean} value The active state.
+   * Attempts to set the selected state of the list item element and it's visual indicators
+   * @param value The value to compare to the parent list element's selected value
+   * @returns Returns whether the list item is selected, or null otherwise.
    */
-  public setSelected(value: boolean): void {
-    if (value) {
-      addClass(LIST_ITEM_CONSTANTS.classes.SELECTED, this._listItemElement);
-      // We are treating selected and activated as the same state, and mdc-states hooks right into --activated
-      // addClass(LIST_ITEM_CONSTANTS.classes.ACTIVATED, this._listItemElement);
-    } else {
-      removeClass(LIST_ITEM_CONSTANTS.classes.SELECTED, this._listItemElement);
-      // removeClass(LIST_ITEM_CONSTANTS.classes.ACTIVATED, this._listItemElement);
+  public trySelect(value: unknown): boolean | null {
+    const list = this._getParentList();
+    if (list?.selectedValue === undefined) {
+      return null;
     }
+
+    const listValues = Array.isArray(list.selectedValue) ? list.selectedValue : [list.selectedValue];
+    const isSelected = listValues.some(v => isDeepEqual(v, value));
+
+    this.tryToggleSelectionControl(isSelected);
+    return isSelected;
   }
 
   /**
-   * Attemps to toggle a checkbox or radio button within the list item if it can find one.
+   * Attempts to toggle a checkbox or radio button within the list item if it can find one.
    */
-  public tryToggleCheckboxRadio(value?: boolean): void {
+  public tryToggleSelectionControl(value?: boolean): void {
     const checkable = this._component.querySelector(LIST_ITEM_CONSTANTS.selectors.CHECKBOX_RADIO_SELECTOR) as HTMLInputElement;
     if (checkable) {
       const force = typeof value === 'boolean';
@@ -165,66 +124,72 @@ export class ListItemAdapter extends BaseAdapter<IListItemComponent> implements 
       // Check if we are just toggling or forcing to a specific checked state
       checkable.checked = force ? value as boolean : !checkable.checked;
 
+      // TODO: This seems weird... Should we be dispatching events on slotted elements when changes are made programmatically?
       if (!force || currentState !== value) {
         checkable.dispatchEvent(new Event('change', { bubbles: true }));
       }
     }
   }
 
-  /** Attempts to set focus to this list item. */
+  public isFocused(): boolean {
+    return this._component.matches(':focus');
+  }
+
   public setFocus(): void {
-    this._listItemElement.focus();
+    this._rootElement.focus({ preventScroll: true });
   }
 
-  /** Returns whether the component has focus or not. */
-  public hasFocus(): boolean {
-    return document.activeElement === this._component || !!this._component.shadowRoot?.activeElement;
+  public animateStateLayer(): void {
+    this._stateLayerElement.playAnimation();
   }
 
-  /**
-   * Sets the role on the list item element.
-   * @param role The role.
-   */
-  public setRole(role: string): void {
-    this._listItemElement.setAttribute('role', role);
+  private _createAnchorRootElement(target: string): HTMLAnchorElement {
+    const a = document.createElement('a');
+    a.classList.add(LIST_ITEM_CONSTANTS.classes.ROOT);
+    a.setAttribute('part', 'root');
+    a.target = target;
+    return a;
   }
 
-  public setDisabled(disabled: boolean): void {
-    toggleClass(this._listItemElement, disabled, LIST_ITEM_CONSTANTS.classes.DISABLED);
+  private _createDefaultRootElement(): HTMLElement {
+    const div = document.createElement('div');
+    div.classList.add(LIST_ITEM_CONSTANTS.classes.ROOT);
+    div.setAttribute('part', 'root');
+    div.tabIndex = this._component.nonInteractive || this._component.disabled ? -1 : 0;
+    return div;
   }
 
-  public setDense(dense: boolean): void {
-    toggleClass(this._listItemElement, dense, LIST_ITEM_CONSTANTS.classes.DENSE);
+  private _getParentList(): IListComponent | null {
+    return this._component.closest(LIST_CONSTANTS.elementName);
   }
 
-  public setIndented(indented: boolean): void {
-    toggleClass(this._listItemElement, indented, LIST_ITEM_CONSTANTS.classes.INDENTED);
-  }
-
-  public setWrap(value: boolean): void {
-    toggleClass(this._listItemElement, value, LIST_ITEM_CONSTANTS.classes.WRAP);
-  }
-
-  /**
-   * Attempts to set the selected state of the list item element and it's visual indicators
-   * @param value The value to compare to the parent list element's selected value
-   * @returns Returns whether the list item is selected or not
-   */
-  public trySelect(value: unknown): boolean | null {
-    const list = requireParent<IListComponent>(this._component, LIST_CONSTANTS.elementName);
-    if (!list || list.selectedValue === undefined) {
-      return null;
+  private _inheritParentListProps(list: IListComponent): void {
+    if (list.hasAttribute(LIST_CONSTANTS.attributes.NON_INTERACTIVE) || list.hasAttribute(LIST_CONSTANTS.attributes.STATIC)) {
+      this._component.nonInteractive = true;
     }
-
-    const listValues = list.selectedValue instanceof Array ? list.selectedValue : [list.selectedValue];
-    const isSelected = listValues.some(v => isDeepEqual(v, value));
-
-    this.setSelected(isSelected);
-    this.tryToggleCheckboxRadio(isSelected);
-    return isSelected;
+    if (list.hasAttribute(LIST_CONSTANTS.attributes.DISABLED)) {
+      this._component.disabled = true;
+    }
+    if (list.hasAttribute(LIST_CONSTANTS.attributes.DENSE)) {
+      this._component.dense = true;
+    }
+    if (list.getAttribute(LIST_CONSTANTS.attributes.PROPAGATE_CLICK) === 'false') {
+      this._component.propagateClick = false;
+    }
+    if (list.hasAttribute(LIST_CONSTANTS.attributes.INDENTED)) {
+      this._component.indented = true;
+    }
+    if (list.hasAttribute(LIST_CONSTANTS.attributes.TWO_LINE)) {
+      this._component.twoLine = true;
+    }
+    if (list.hasAttribute(LIST_CONSTANTS.attributes.THREE_LINE)) {
+      this._component.threeLine = true;
+    }
   }
 
-  public userInteractionListener(): ReturnType<typeof createUserInteractionListener> {
-    return createUserInteractionListener(this._listItemElement);
+  private _setRole(list: IListComponent | null): void {
+    const listRole = list?.getAttribute('role');
+    const role = ListComponentItemRole[listRole as string] ?? 'listitem';
+    this._component.setAttribute('role', role);
   }
 }
