+++ src/lib/list/list-item/list-item.ts
@@ -1,30 +1,29 @@
-import { CustomElement, attachShadowTemplate, requireParent, elementParents, coerceBoolean, FoundationProperty } from '@tylertech/forge-core';
+import { CustomElement, attachShadowTemplate, ICustomElement, FoundationProperty, coerceBoolean } from '@tylertech/forge-core';
 import { ListItemAdapter } from './list-item-adapter';
 import { ListItemFoundation } from './list-item-foundation';
 import { IListItemSelectEventData, LIST_ITEM_CONSTANTS } from './list-item-constants';
-import { LIST_CONSTANTS } from '../list/list-constants';
-import { IListComponent } from '../list';
-import { BaseComponent, IBaseComponent } from '../../core/base/base-component';
+import { StateLayerComponent } from '../../state-layer';
+import { FocusIndicatorComponent } from '../../focus-indicator';
 
-import defaultTemplate from './list-item.html';
+import template from './list-item.html';
 import styles from './list-item.scss';
 
-export interface IListItemComponent extends IBaseComponent {
-  static: boolean;
-  twoLine: boolean;
-  threeLine: boolean;
-  active: boolean;
-  selected: boolean;
-  value: any;
+export interface IListItemComponent extends ICustomElement {
   href: string;
   target: string;
-  ripple: boolean;
+  /** @deprecated Use nonInteractive instead. */
+  static: boolean;
+  nonInteractive: boolean;
   disabled: boolean;
+  selected: boolean;
+  active: boolean;
+  value: unknown;
   dense: boolean;
   propagateClick: boolean;
   indented: boolean;
+  twoLine: boolean;
+  threeLine: boolean;
   wrap: boolean;
-  focus(): void;
 }
 
 declare global {
@@ -38,30 +37,121 @@ declare global {
 }
 
 /**
- * The custom element class behind the `<forge-list-item>` element.
- * 
  * @tag forge-list-item
+ * 
+ * @summary List items are individual rows of content inside of a list.
+ * 
+ * @property {string} href - The href of the list item. This forces the list item to render as an anchor element.
+ * @property {string} target - The target of the list item when an `href` is set. Defaults to `'_blank'`.
+ * @property {boolean} nonInteractive - If true, the list item will not be interactive.
+ * @property {boolean} static - If true, the list item will not be interactive. Deprecated use `nonInteractive` instead.
+ * @property {boolean} disabled - Disables the list item.
+ * @property {boolean} selected - Applies the selected state to the list item.
+ * @property {boolean} active - Applies the active state to the list item by emulating its focused state.
+ * @property {unknown} value - The unique value of the list item.
+ * @property {boolean} dense - Applies the dense state to the list item.
+ * @property {boolean} propagateClick - If true, the list item will not propagate click events to itself and therefore cannot receive focus.
+ * @property {boolean} indented - Applies the indented state by adding margin to the start of the list item.
+ * @property {boolean} twoLine - Sets the list item height to support at least two lines of text.
+ * @property {boolean} threeLine - Sets the list item height to support at least three lines of text.
+ * @property {boolean} wrap - Sets the list item to wrap its text content.
+ * 
+ * @attribute {string} href - The href of the list item. This forces the list item to render as an anchor element.
+ * @attribute {string} target - The target of the list item when an `href` is set. Defaults to `'_blank'`.
+ * @attribute {boolean} non-interactive - If true, the list item will not be interactive.
+ * @attribute {boolean} static - If true, the list item will not be interactive. Deprecated use `non-interactive` instead.
+ * @attribute {boolean} disabled - Disables the list item.
+ * @attribute {boolean} selected - Applies the selected state to the list item.
+ * @attribute {boolean} active - Applies the active state to the list item by emulating its focused state.
+ * @attribute {unknown} value - The unique value of the list item.
+ * @attribute {boolean} dense - Applies the dense state to the list item.
+ * @attribute {boolean} propagate-click - If applied, the list item will not propagate click events to itself and therefore cannot receive focus.
+ * @attribute {boolean} indented - Applies the indented state by adding margin to the start of the list item.
+ * @attribute {boolean} two-line - Sets the list item height to support at least two lines of text.
+ * @attribute {boolean} three-line - Sets the list item height to support at least three lines of text.
+ * @attribute {boolean} wrap - Sets the list item to wrap its text content.
+ * 
+ * @event {CustomEvent<IListItemSelectEventData>} forge-list-item-select - Fires when the list item is selected.
+ * 
+ * @slot - The primary text.
+ * @slot primary-text - The primary text. A named alias for the default slot.
+ * @slot secondary-text - The secondary text.
+ * @slot tertiary-text - The tertiary text.
+ * @slot title - The title element. An alias for the primary-text slot for backwards compatibility.
+ * @slot subtitle - The subtitle element. An alias for the secondary-text slot for backwards compatibility.
+ * @slot tertiary-title - The tertiary title element. An alias for the tertiary-text slot for backwards compatibility.
+ * @slot leading - The leading content.
+ * @slot trailing - The trailing element.
+ * @slot avatar - The avatar content.
+ * 
+ * @csspart root - The root container element.
+ * @csspart text-container - The container for the text content.
+ * @csspart focus-indicator__indicator - The forwarded focus indicator's internal indicator element.
+ * @csspart state-layer__surface - The forwarded state layer's internal surface element.
+ * 
+ * @cssprop --forge-list-item-background-color - The background color.
+ * @cssprop --forge-list-item-shape - The shape of the list item.
+ * @cssprop --forge-list-item-padding - The padding inside of the container element.
+ * @cssprop --forge-list-item-margin - The margin around the host element.
+ * @cssprop --forge-list-item-height - The height of the container.
+ * @cssprop --forge-list-item-dense-height - The height when in the dense state.
+ * @cssprop --forge-list-item-indent - The margin inline state when in the indented state.
+ * @cssprop --forge-list-item-supporting-text-color - The text color of the supporting text.
+ * @cssprop --forge-list-item-supporting-line-height - The line height of the supporting text.
+ * @cssprop --forge-list-item-selected-color - The color when in the selected state.
+ * @cssprop --forge-list-item-opacity - The opacity of the background color when in the disabled state.
+ * @cssprop --forge-list-item-selected-leading-color - The color of the leading content when in the selected state.
+ * @cssprop --forge-list-item-selected-trailing-color - The color of the trailing content when in the selected state.
+ * @cssprop --forge-list-item-selected-supporting-text-color - The color of the supporting text when in the selected state.
+ * @cssprop --forge-list-item-disabled-opacity - The opacity of the element when in the disabled state.
+ * @cssprop --forge-list-item-disabled-cursor - The cursor when in the disabled state.
+ * @cssprop --forge-list-item-one-line-height - The line height when in the one/single line state.
+ * @cssprop --forge-list-item-two-line-height - The line height when in the two line state.
+ * @cssprop --forge-list-item-three-line-height - The line height when in the three line state.
+ * @cssprop --forge-list-item-dense-one-line-height - The line height when in the dense one/single line state.
+ * @cssprop --forge-list-item-dense-two-line-height - The line height when in the dense two line state.
+ * @cssprop --forge-list-item-dense-three-line-height - The line height when in the dense three line state.
+ * @cssprop --forge-list-item-dense-font-size - The font size when in the dense state.
+ * @cssprop --forge-list-item-dense-indent - The margin inline state when in the dense indented state.
+ * @cssprop --forge-list-item-dense-leading-margin-end - The margin end of the leading content when in the dense state.
+ * @cssprop --forge-list-item-dense-trailing-margin-start - The margin start of the trailing content when in the dense state.
+ * @cssprop --forge-list-item-leading-margin-start - The margin start of the leading content.
+ * @cssprop --forge-list-item-leading-margin-end - The margin end of the leading content.
+ * @cssprop --forge-list-item-leading-selected-color - The color of the leading content when in the selected state.
+ * @cssprop --forge-list-item-trailing-margin-start - The margin start of the trailing content.
+ * @cssprop --forge-list-item-trailing-margin-end - The margin end of the trailing content.
+ * @cssprop --forge-list-item-trailing-selected-color - The color of the trailing content when in the selected state.
+ * @cssprop --forge-list-item-avatar-background-color - The background color of the avatar container.
+ * @cssprop --forge-list-item-avatar-color - The foreground color of the avatar container.
+ * @cssprop --forge-list-item-avatar-shape - The shape of the avatar container.
+ * @cssprop --forge-list-item-avatar-margin-start - The margin start of the avatar container.
+ * @cssprop --forge-list-item-avatar-margin-end - The margin end of the avatar container.
+ * @cssprop --forge-list-item-avatar-size - The height & width of the avatar container.
  */
 @CustomElement({
-  name: LIST_ITEM_CONSTANTS.elementName
+  name: LIST_ITEM_CONSTANTS.elementName,
+  dependencies: [
+    StateLayerComponent,
+    FocusIndicatorComponent
+  ]
 })
-export class ListItemComponent extends BaseComponent implements IListItemComponent {
+export class ListItemComponent extends HTMLElement implements IListItemComponent {
   public static get observedAttributes(): string[] {
     return [
-      LIST_ITEM_CONSTANTS.attributes.STATIC,
-      LIST_ITEM_CONSTANTS.attributes.TWO_LINE,
-      LIST_ITEM_CONSTANTS.attributes.THREE_LINE,
-      LIST_ITEM_CONSTANTS.attributes.ACTIVE,
-      LIST_ITEM_CONSTANTS.attributes.SELECTED,
-      LIST_ITEM_CONSTANTS.attributes.VALUE,
-      LIST_ITEM_CONSTANTS.attributes.HREF,
-      LIST_ITEM_CONSTANTS.attributes.TARGET,
-      LIST_ITEM_CONSTANTS.attributes.RIPPLE,
-      LIST_ITEM_CONSTANTS.attributes.DISABLED,
-      LIST_ITEM_CONSTANTS.attributes.DENSE,
-      LIST_ITEM_CONSTANTS.attributes.PROPAGATE_CLICK,
-      LIST_ITEM_CONSTANTS.attributes.INDENTED,
-      LIST_ITEM_CONSTANTS.attributes.WRAP
+      LIST_ITEM_CONSTANTS.observedAttributes.HREF,
+      LIST_ITEM_CONSTANTS.observedAttributes.TARGET,
+      LIST_ITEM_CONSTANTS.observedAttributes.STATIC,
+      LIST_ITEM_CONSTANTS.observedAttributes.NON_INTERACTIVE,
+      LIST_ITEM_CONSTANTS.observedAttributes.DISABLED,
+      LIST_ITEM_CONSTANTS.observedAttributes.SELECTED,
+      LIST_ITEM_CONSTANTS.observedAttributes.ACTIVE,
+      LIST_ITEM_CONSTANTS.observedAttributes.VALUE,
+      LIST_ITEM_CONSTANTS.observedAttributes.DENSE,
+      LIST_ITEM_CONSTANTS.observedAttributes.PROPAGATE_CLICK,
+      LIST_ITEM_CONSTANTS.observedAttributes.INDENTED,
+      LIST_ITEM_CONSTANTS.observedAttributes.TWO_LINE,
+      LIST_ITEM_CONSTANTS.observedAttributes.THREE_LINE,
+      LIST_ITEM_CONSTANTS.observedAttributes.WRAP
     ];
   }
 
@@ -69,23 +159,11 @@ export class ListItemComponent extends BaseComponent implements IListItemCompone
 
   constructor() {
     super();
-    attachShadowTemplate(this, defaultTemplate, styles);
+    attachShadowTemplate(this, template, styles);
     this._foundation = new ListItemFoundation(new ListItemAdapter(this));
   }
 
   public connectedCallback(): void {
-    // To simulate the :host-context() selector for Firefox until they implement it, we need to determine if the
-    // list item is within a drawer for auto-styling the list item when included within a drawer. Check to see if
-    // any of the parents of this element are a drawer.
-    if (!this.hasAttribute(LIST_ITEM_CONSTANTS.attributes.DRAWER_CONTEXT) && elementParents(this).some(el => ['forge-drawer', 'forge-modal-drawer', 'forge-mini-drawer'].includes(el.tagName.toLowerCase()))) {
-      this.setAttribute(LIST_ITEM_CONSTANTS.attributes.DRAWER_CONTEXT, 'true');
-    }
-
-    const list = requireParent<IListComponent>(this, LIST_CONSTANTS.elementName);
-    if (list) {
-      this._inheritParentListProps(list);
-    }
-
     this._foundation.initialize();
   }
 
@@ -95,124 +173,93 @@ export class ListItemComponent extends BaseComponent implements IListItemCompone
 
   public attributeChangedCallback(name: string, oldValue: string, newValue: string): void {
     switch (name) {
-      case LIST_ITEM_CONSTANTS.attributes.STATIC:
-        this.static = coerceBoolean(newValue);
-        break;
-      case LIST_ITEM_CONSTANTS.attributes.TWO_LINE:
-        this.twoLine = coerceBoolean(newValue);
-        break;
-      case LIST_ITEM_CONSTANTS.attributes.THREE_LINE:
-        this.threeLine = coerceBoolean(newValue);
-        break;
-      case LIST_ITEM_CONSTANTS.attributes.ACTIVE:
-        this.active = coerceBoolean(newValue);
-        break;
-      case LIST_ITEM_CONSTANTS.attributes.SELECTED:
-        this.selected = coerceBoolean(newValue);
-        break;
-      case LIST_ITEM_CONSTANTS.attributes.VALUE:
-        this.value = newValue;
-        break;
-      case LIST_ITEM_CONSTANTS.attributes.HREF:
+      case LIST_ITEM_CONSTANTS.observedAttributes.HREF:
         this.href = newValue;
         break;
-      case LIST_ITEM_CONSTANTS.attributes.TARGET:
+      case LIST_ITEM_CONSTANTS.observedAttributes.TARGET:
         this.target = newValue;
         break;
-      case LIST_ITEM_CONSTANTS.attributes.RIPPLE:
-        this.ripple = coerceBoolean(newValue);
+      case LIST_ITEM_CONSTANTS.observedAttributes.NON_INTERACTIVE:
+      case LIST_ITEM_CONSTANTS.observedAttributes.STATIC:
+        this.nonInteractive = coerceBoolean(newValue);
         break;
-      case LIST_ITEM_CONSTANTS.attributes.DISABLED:
+      case LIST_ITEM_CONSTANTS.observedAttributes.DISABLED:
         this.disabled = coerceBoolean(newValue);
         break;
-      case LIST_ITEM_CONSTANTS.attributes.DENSE:
+      case LIST_ITEM_CONSTANTS.observedAttributes.SELECTED:
+        this.selected = coerceBoolean(newValue);
+        break;
+      case LIST_ITEM_CONSTANTS.observedAttributes.ACTIVE:
+        this.active = coerceBoolean(newValue);
+        break;
+      case LIST_ITEM_CONSTANTS.observedAttributes.VALUE:
+        this.value = newValue;
+        break;
+      case LIST_ITEM_CONSTANTS.observedAttributes.DENSE:
         this.dense = coerceBoolean(newValue);
         break;
-      case LIST_ITEM_CONSTANTS.attributes.PROPAGATE_CLICK:
+      case LIST_ITEM_CONSTANTS.observedAttributes.PROPAGATE_CLICK:
         this.propagateClick = coerceBoolean(newValue);
         break;
-      case LIST_ITEM_CONSTANTS.attributes.INDENTED:
+      case LIST_ITEM_CONSTANTS.observedAttributes.INDENTED:
         this.indented = coerceBoolean(newValue);
         break;
-      case LIST_ITEM_CONSTANTS.attributes.WRAP:
+      case LIST_ITEM_CONSTANTS.observedAttributes.TWO_LINE:
+        this.twoLine = coerceBoolean(newValue);
+        break;
+      case LIST_ITEM_CONSTANTS.observedAttributes.THREE_LINE:
+        this.threeLine = coerceBoolean(newValue);
+        break;
+      case LIST_ITEM_CONSTANTS.observedAttributes.WRAP:
         this.wrap = coerceBoolean(newValue);
         break;
     }
   }
 
-  private _inheritParentListProps(list: IListComponent): void {
-    if (list.hasAttribute(LIST_CONSTANTS.attributes.STATIC)) {
-      this.static = true;
-    }
-    if (list.hasAttribute(LIST_CONSTANTS.attributes.DENSE)) {
-      this.dense = true;
-    }
-    if (list.getAttribute(LIST_CONSTANTS.attributes.PROPAGATE_CLICK) === 'false') {
-      this.propagateClick = false;
-    }
-    if (list.hasAttribute(LIST_CONSTANTS.attributes.INDENTED)) {
-      this.indented = true;
-    }
-  }
-
-  /** Gets/sets whether the static state of this list item. */
-  @FoundationProperty()
-  public declare static: boolean;
-
-  /** Gets/sets whether the list item displays two lines of text. */
-  @FoundationProperty()
-  public declare twoLine: boolean;
-
-  /** Gets/sets whether the list item displays three lines of text. */
-  @FoundationProperty()
-  public declare threeLine: boolean;
-
-  /** Gets/sets whether the list item is active or not. */
-  @FoundationProperty()
-  public declare active: boolean;
-
-  /** Gets/sets whether the list item is selected or not. */
-  @FoundationProperty()
-  public declare selected: boolean;
-
-  /** Gets/sets list item value. */
-  @FoundationProperty()
-  public declare value: any;
-
-  /** Gets/sets the href link that this list item will send the browser to when clicked. */
-  @FoundationProperty()
-  public declare href: string;
-
-  /** Gets/sets the href link target. Only pertains when `href` is also used. */
-  @FoundationProperty()
-  public declare target: string;
-
-  /** Gets/sets whether the list item has a ripple or not. */
-  @FoundationProperty()
-  public declare ripple: boolean;
-
-  /** Gets/sets whether the list item is disabled or not. */
-  @FoundationProperty()
-  public declare disabled: boolean;
-
-  /** Gets/sets whether the list item is using dense styles or not. */
-  @FoundationProperty()
-  public declare dense: boolean;
-
-  /** Gets/sets whether the list item allows mousedown events through to the underlying list item element. Default is true. */
-  @FoundationProperty()
-  public declare propagateClick: boolean;
-
-  /** Gets/sets whether the list item is indented or not. Default is false. */
-  @FoundationProperty()
-  public declare indented: boolean;
-
-  /** Gets/sets whether the list item content is wrapped or not. Default is true. */
-  @FoundationProperty()
-  public declare wrap: boolean;
-
-  /** Sets focus to this list item. */
   public override focus(): void {
     this._foundation.setFocus();
   }
+
+  @FoundationProperty()
+  public declare href: string;
+
+  @FoundationProperty()
+  public declare target: string;
+
+  /** @deprecated Use nonInteractive instead. */
+  @FoundationProperty()
+  public declare static: boolean;
+
+  @FoundationProperty()
+  public declare nonInteractive: boolean;
+
+  @FoundationProperty()
+  public declare disabled: boolean;
+
+  @FoundationProperty()
+  public declare selected: boolean;
+
+  @FoundationProperty()
+  public declare active: boolean;
+
+  @FoundationProperty()
+  public declare value: unknown;
+
+  @FoundationProperty()
+  public declare dense: boolean;
+
+  @FoundationProperty()
+  public declare propagateClick: boolean;
+
+  @FoundationProperty()
+  public declare indented: boolean;
+
+  @FoundationProperty()
+  public declare twoLine: boolean;
+
+  @FoundationProperty()
+  public declare threeLine: boolean;
+
+  @FoundationProperty()
+  public declare wrap: boolean;
 }
